renderer: use size_t for expression numOps

+++ src/engine/renderer/tr_local.h
@@ -940,7 +940,7 @@ enum class realtimeLightingRenderer_t { LEGACY, TILED };
 	struct expression_t
 	{
 		expOperation_t ops[ MAX_EXPRESSION_OPS ];
-		uint8_t numOps;
+		size_t numOps;
 	};
 
 	struct waveForm_t
